+++ pages/add-viewer.tsx
@@ -69,7 +69,7 @@ const AddViewerToProject: React.FC = () => {
 
   useEffect(() => {
     if (user === undefined) {
-      push("/");
+      push("/login");
       return;
     }
