+++ pages/subscription.tsx
@@ -80,7 +80,7 @@ const Payment: React.FC = () => {
 
   useEffect(() => {
     if (user === undefined) {
-      push("/");
+      push("/login");
       return;
     }
